pyproject.toml
[project]
name = "Postgres MCP Server"
version = "0.1.0"
description = "An MCP server to connect and read data from Posgres databases"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.23.3",
    "psycopg2-binary>=2.9.11",
]
